requirements.txt
pytorch_lighting
omegaconf
tensorboardX